+++ packages/ws/src/lib/mocha.ts
@@ -1,9 +1,0 @@
-import Mocha from 'mocha';
-
-export function testAsync(files: string[]) {
-  const mocha = new Mocha();
-  files.forEach((file) => mocha.addFile(file));
-  return new Promise<number>((resolve) => {
-    mocha.run(resolve);
-  });
-}