+++ packages/ws/src/lib/webpack/node.ts
@@ -7,20 +7,16 @@ import {
 } from './options';
 import { EnvOptions } from '../../options';
 
-export const getNodeBuildConfig = async (
-  options: EnvOptions
-): Promise<WebpackConfig> => ({
+export const getNodeBuildConfig = (options: EnvOptions): WebpackConfig => ({
   ...baseConfig,
   ...nodeConfig,
-  ...(await getEntryAndOutput('node', 'build')),
+  ...getEntryAndOutput('node', 'build'),
   ...getModuleAndPlugins('node', 'build', options)
 });
 
-export const getNodeUnitConfig = async (
-  options: EnvOptions
-): Promise<WebpackConfig> => ({
+export const getNodeUnitConfig = (options: EnvOptions): WebpackConfig => ({
   ...baseConfig,
   ...nodeConfig,
-  ...(await getEntryAndOutput('node', 'unit')),
+  ...getEntryAndOutput('node', 'unit'),
   ...getModuleAndPlugins('node', 'unit', options)
 });