+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 The changelog is currently handwritten.
 
+# HEAD
+
+- Added Cypress example.
+
 # 3.0.0-2 (2018-07-27)
 
 - Fixed watch config and added support for history fallback, so reloading on any URL works correctly.